+++ ui/app/components/list-view.js
@@ -2,9 +2,8 @@ import Ember from 'ember';
 import { pluralize } from 'ember-inflector';
 import { task } from 'ember-concurrency';
 
-const { computed, inject } = Ember;
+const { computed } = Ember;
 export default Ember.Component.extend({
-  flashMessages: inject.service(),
   tagName: '',
   items: null,
   itemNoun: 'item',
@@ -12,18 +11,5 @@ export default Ember.Component.extend({
   emptyMessage: computed('itemNoun', function() {
     let items = pluralize(this.get('itemNoun'));
     return `There are currently no ${items}`;
-  }),
-
-  saveItem: task(function*() {}),
-  deleteItem: task(function*(model, successMessage, failureMessage) {
-    let flash = this.get('flashMessages');
-    try {
-      yield model.destroyRecord();
-      flash.success(successMessage);
-    } catch (e) {
-      let errString = e.errors.join(',');
-      flash.danger(failureMessage + errString);
-      model.rollbackAttributes();
-    }
   }),
 });
