+++ ui/app/serializers/namespace.js
@@ -2,11 +2,11 @@ import ApplicationSerializer from './application';
 
 export default ApplicationSerializer.extend({
   normalizeList(payload) {
-    let pk = 'path';
     const data = payload.data.keys
       ? payload.data.keys.map(key => ({
-          [pk]: key,
-          id: key.slice(0, -1),
+          path: key,
+          // remove the trailing slash from the id
+          id: key.replace(/\/$/, ''),
         }))
       : payload.data;
 
@@ -15,8 +15,10 @@ export default ApplicationSerializer.extend({
 
   normalizeResponse(store, primaryModelClass, payload, id, requestType) {
     const nullResponses = ['deleteRecord', 'createRecord'];
-    let cid = id || payload.id;
-    let normalizedPayload = nullResponses.includes(requestType) ? { id: cid } : this.normalizeList(payload);
+    let cid = (id || payload.id || '').replace(/\/$/, '');
+    let normalizedPayload = nullResponses.includes(requestType)
+      ? { id: cid, path: cid }
+      : this.normalizeList(payload);
     return this._super(store, primaryModelClass, normalizedPayload, id, requestType);
   },
 });
