+++ ui/app/templates/vault/cluster/access/namespaces.hbs
@@ -1,2 +1,1 @@
-
 {{outlet}}
